sqLite Data

+++ app/src/main/java/com/school/apicallskt/room_db_kotlin/model/CourseRoomKtModel.kt
@@ -11,5 +11,5 @@ data class CourseRoomKtModel(
     var courseDescription: String? = null
 ) {
     @PrimaryKey(autoGenerate = true)
-    var id: Long = 0
+    var id: Int = 0
 }
